requirements.txt
asgiref==3.8.1
Django==4.1.13
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
pymongo==3.12.3
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.2.4
tzdata==2024.1
channels==3.0.5
drf-yasg==1.21.7